move public collator interface into its own package
