+++ miner/default_collator.go
@@ -2,13 +2,11 @@ package miner
 
 import (
 	"context"
-	"errors"
 	"sync"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
+	"github.com/ethereum/go-ethereum/miner/collator"
 )
 
 const (
@@ -33,9 +31,9 @@ type DefaultCollator struct {
 	recommitMu   sync.Mutex
 	recommit     time.Duration
 	minRecommit  time.Duration
-	miner        MinerState
+	miner        collator.MinerState
 	exitCh       <-chan struct{}
-	pool         Pool
+	pool         collator.Pool
 	resubmitHook func(time.Duration, time.Duration) // Method to call upon updating resubmitting interval.
 }
 
@@ -84,102 +82,7 @@ func (c *DefaultCollator) adjustRecommit(ratio float64, inc bool) {
 	}
 }
 
-func submitTransactions(ctx context.Context, bs BlockState, txs *types.TransactionsByPriceAndNonce, timer *time.Timer) uint {
-	header := bs.Header()
-	availableGas := header.GasLimit
-	var numTxsAdded uint = 0
-
-	for {
-		select {
-		case <-ctx.Done():
-			return numTxsAdded
-		default:
-		}
-
-		if timer != nil {
-			select {
-			case <-timer.C:
-				return numTxsAdded
-			default:
-			}
-		}
-
-		// Retrieve the next transaction and abort if all done
-		tx := txs.Peek()
-		if tx == nil {
-			break
-		}
-		// Enough space for this tx?
-		if availableGas < tx.Gas() {
-			txs.Pop()
-			continue
-		}
-		from, _ := types.Sender(bs.Signer(), tx)
-
-		receipt, err := bs.AddTransaction(tx)
-		switch {
-		case errors.Is(err, ErrGasLimitReached):
-			// Pop the current out-of-gas transaction without shifting in the next from the account
-			log.Trace("Gas limit exceeded for current block", "sender", from)
-			txs.Pop()
-
-		case errors.Is(err, ErrNonceTooLow):
-			// New head notification data race between the transaction pool and miner, shift
-			log.Trace("Skipping transaction with low nonce", "sender", from, "nonce", tx.Nonce())
-			txs.Shift()
-
-		case errors.Is(err, ErrNonceTooHigh):
-			// Reorg notification data race between the transaction pool and miner, skip account =
-			log.Trace("Skipping account with high nonce", "sender", from, "nonce", tx.Nonce())
-			txs.Pop()
-
-		case errors.Is(err, nil):
-			availableGas = header.GasLimit - receipt.CumulativeGasUsed
-
-			numTxsAdded++
-			// Everything ok, collect the logs and shift in the next transaction from the same account
-			txs.Shift()
-
-		case errors.Is(err, ErrTxTypeNotSupported):
-			// Pop the unsupported transaction without shifting in the next from the account
-			log.Trace("Skipping unsupported transaction type", "sender", from, "type", tx.Type())
-			txs.Pop()
-		default:
-			// Strange error, discard the transaction and get the next in line (note, the
-			// nonce-too-high clause will prevent us from executing in vain).
-			log.Debug("Transaction failed, account skipped", "hash", tx.Hash(), "err", err)
-			txs.Shift()
-		}
-	}
-
-	return numTxsAdded
-}
-
-func FillTransactions(ctx context.Context, bs BlockState, timer *time.Timer, pendingTxs map[common.Address]types.Transactions, locals []common.Address) uint {
-	header := bs.Header()
-	if len(pendingTxs) == 0 {
-		return 0
-	}
-	// Split the pending transactions into locals and remotes
-	localTxs, remoteTxs := make(map[common.Address]types.Transactions), pendingTxs
-	for _, account := range locals {
-		if accountTxs := remoteTxs[account]; len(accountTxs) > 0 {
-			delete(remoteTxs, account)
-			localTxs[account] = accountTxs
-		}
-	}
-	var txCount uint = 0
-	if len(localTxs) > 0 {
-		txCount += submitTransactions(ctx, bs, types.NewTransactionsByPriceAndNonce(bs.Signer(), localTxs, header.BaseFee), timer)
-	}
-	if len(remoteTxs) > 0 {
-		txCount += submitTransactions(ctx, bs, types.NewTransactionsByPriceAndNonce(bs.Signer(), remoteTxs, header.BaseFee), timer)
-	}
-
-	return txCount
-}
-
-func (c *DefaultCollator) workCycle(work BlockCollatorWork) {
+func (c *DefaultCollator) workCycle(work collator.BlockCollatorWork) {
 	ctx := work.Ctx
 	emptyBs := work.Block
 
@@ -192,7 +95,7 @@ func (c *DefaultCollator) workCycle(work BlockCollatorWork) {
 		bs := emptyBs.Copy()
 		pendingTxs, _ := c.pool.Pending(true)
 		locals := c.pool.Locals()
-		FillTransactions(ctx, bs, timer, pendingTxs, locals)
+		collator.FillTransactions(ctx, bs, timer, pendingTxs, locals)
 		bs.Commit()
 		shouldContinue := false
 
@@ -254,7 +157,7 @@ func (c *DefaultCollator) SetRecommit(interval time.Duration) {
 	}
 }
 
-func (c *DefaultCollator) CollateBlocks(miner MinerState, pool Pool, blockCh <-chan BlockCollatorWork, exitCh <-chan struct{}) {
+func (c *DefaultCollator) CollateBlocks(miner collator.MinerState, pool collator.Pool, blockCh <-chan collator.BlockCollatorWork, exitCh <-chan struct{}) {
 	c.miner = miner
 	c.exitCh = exitCh
 	c.pool = pool
@@ -271,11 +174,11 @@ func (c *DefaultCollator) CollateBlocks(miner MinerState, pool Pool, blockCh <-c
 	}
 }
 
-func (c *DefaultCollator) CollateBlock(bs BlockState, pool Pool) {
+func (c *DefaultCollator) CollateBlock(bs collator.BlockState, pool collator.Pool) {
 	c.pool = pool // TODO weird to set this here
 	pendingTxs, _ := pool.Pending(true)
 	locals := pool.Locals()
-	FillTransactions(context.Background(), bs, nil, pendingTxs, locals)
+	collator.FillTransactions(context.Background(), bs, nil, pendingTxs, locals)
 	bs.Commit()
 }
